Release v1.

+++ lab6notes/app/src/main/java/pl/proenix/android/us2pum/lab6notes/NotesListAdapter.java
@@ -63,7 +63,7 @@ public class NotesListAdapter extends RecyclerView.Adapter<NotesListAdapter.Note
         holder.itemView.setOnClickListener(v -> {
             Bundle bundle = new Bundle();
             bundle.putSerializable("mode", NoteCreateReadUpdateFragment.NoteEditMode.NOTE_UPDATE);
-            bundle.putLong("noteID", (Long) noteList.get(holder.getAdapterPosition()).getID());
+            bundle.putLong("noteID", noteList.get(holder.getAdapterPosition()).getID());
             Navigation.findNavController(v).navigate(R.id.action_notesListFragment_to_noteCreateReadUpdateFragment, bundle);
         });
         // On long click select note.
